requirements.txt
ffmpeg-python
wand
fpdf
